\docType{package}
\name{hyperSpec.matrixStats-package}
\alias{hyperSpec.matrixStats-package}
\title{Package hyperSpec.matrixStats}
\description{
  matrixStats for hyperSpec objects This package provides
  the functions defined in
  \code{\link[matrixStats]{matrixStats-package}} for
  \code{\link[hyperSpec]{hyperSpec-class}} objects.
}
\author{
  C. Beleites

  Maintainer: Claudia Beleites
  <Claudia.Beleites@ipht-jena.de>
}
\seealso{
  \code{citation ("hyperSpec-matrixStats")} produces the
  correct citation.

  \code{package?hyperSpec.matrixStats} for information
  about the package
}
\keyword{package}

