% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{matrixStats}
\alias{anyMissing,hyperSpec-method}
\alias{colAlls.hyperSpec}
\alias{colAnys.hyperSpec}
\alias{colAvgsPerRowSet.hyperSpec}
\alias{colCollapse.hyperSpec}
\alias{colCounts.hyperSpec}
\alias{colDiffs,hyperSpec-method}
\alias{colIQRs,hyperSpec-method}
\alias{colMads,hyperSpec,ANY-method}
\alias{colMaxs,hyperSpec-method}
\alias{colMedians,hyperSpec-method}
\alias{colMins,hyperSpec-method}
\alias{colOrderStats,hyperSpec-method}
\alias{colProds,hyperSpec-method}
\alias{colQuantiles,hyperSpec-method}
\alias{colRanges,hyperSpec-method}
\alias{colRanks,hyperSpec-method}
\alias{colSds,hyperSpec-method}
\alias{colVars,hyperSpec-method}
\alias{colWeightedMeans.hyperSpec}
\alias{colWeightedMedians.hyperSpec}
\alias{matrixStats}
\alias{rowAlls.hyperSpec}
\alias{rowAnys.hyperSpec}
\alias{rowAvgsPerColSet.hyperSpec}
\alias{rowCollapse.hyperSpec}
\alias{rowCounts.hyperSpec}
\alias{rowDiffs,hyperSpec-method}
\alias{rowIQRs,hyperSpec-method}
\alias{rowMads,hyperSpec,ANY-method}
\alias{rowMaxs,hyperSpec-method}
\alias{rowMedians,hyperSpec-method}
\alias{rowMins,hyperSpec-method}
\alias{rowOrderStats,hyperSpec-method}
\alias{rowProds,hyperSpec-method}
\alias{rowQuantiles,hyperSpec-method}
\alias{rowRanges,hyperSpec-method}
\alias{rowRanks,hyperSpec-method}
\alias{rowSds,hyperSpec-method}
\alias{rowVars,hyperSpec-method}
\alias{rowWeightedMeans.hyperSpec}
\alias{rowWeightedMedians.hyperSpec}
\title{matrixStats functions for hyperSpec objects}
\usage{
\S4method{anyMissing}{hyperSpec}(x, ..., na.rm = TRUE)

\method{colAlls}{hyperSpec}(x, value = TRUE, na.rm = TRUE, ...)

\method{colAnys}{hyperSpec}(x, value = TRUE, na.rm = TRUE, ...)

\method{rowAlls}{hyperSpec}(x, value = TRUE, na.rm = TRUE, ...)

\method{rowAnys}{hyperSpec}(x, value = TRUE, na.rm = TRUE, ...)

\method{colCollapse}{hyperSpec}(x, idxs, ..., na.rm = TRUE, label.spc)

\method{colCounts}{hyperSpec}(x, value = TRUE, na.rm = TRUE, ..., label.spc)

\S4method{colDiffs}{hyperSpec}(x, ..., na.rm = TRUE, label.spc)

\S4method{colIQRs}{hyperSpec}(x, ..., na.rm = TRUE, label.spc)

\S4method{colMads}{hyperSpec,ANY}(x, centers = colMedians(x, ...),
  constant = 1.4826, ..., na.rm = TRUE, label.spc)

\S4method{colMaxs}{hyperSpec}(x, na.rm = TRUE, ..., label.spc)

\S4method{colMedians}{hyperSpec}(x, na.rm = TRUE, ..., label.spc)

\S4method{colMins}{hyperSpec}(x, na.rm = TRUE, ..., label.spc)

\S4method{colOrderStats}{hyperSpec}(x, which, ..., na.rm = TRUE, label.spc)

\S4method{colProds}{hyperSpec}(x, na.rm = TRUE, method = c("expSumLog",
  "direct"), ..., label.spc)

\S4method{colQuantiles}{hyperSpec}(x, ..., na.rm = TRUE, label.spc)

\S4method{colRanges}{hyperSpec}(x, na.rm = TRUE, ..., label.spc)

\S4method{colRanks}{hyperSpec}(x, ties.method = c("max", "average", "min"),
  preserveShape = FALSE, ..., na.rm = TRUE, label.spc)

\S4method{colSds}{hyperSpec}(x, ..., na.rm = TRUE, label.spc)

\S4method{colVars}{hyperSpec}(x, na.rm = TRUE, center = NULL, ...,
  label.spc)

\method{colWeightedMeans}{hyperSpec}(x, w = NULL, na.rm = TRUE, ...,
  label.spc)

\method{colWeightedMedians}{hyperSpec}(x, w = NULL, na.rm = TRUE, ...,
  label.spc)

\method{rowCollapse}{hyperSpec}(x, idxs, ..., na.rm = TRUE, label.wavelength)

\method{rowCounts}{hyperSpec}(x, value = TRUE, na.rm = TRUE, ...,
  label.wavelength)

\S4method{rowDiffs}{hyperSpec}(x, ..., na.rm = TRUE, label.wavelength)

\S4method{rowIQRs}{hyperSpec}(x, ..., na.rm = TRUE, label.wavelength)

\S4method{rowMads}{hyperSpec,ANY}(x, centers = rowMedians(x, ...),
  constant = 1.4826, ..., na.rm = TRUE, label.wavelength)

\S4method{rowMaxs}{hyperSpec}(x, na.rm = TRUE, ..., label.wavelength)

\S4method{rowMedians}{hyperSpec}(x, na.rm = TRUE, ..., label.wavelength)

\S4method{rowMins}{hyperSpec}(x, na.rm = TRUE, ..., label.wavelength)

\S4method{rowOrderStats}{hyperSpec}(x, which, ..., na.rm = TRUE,
  label.wavelength)

\S4method{rowProds}{hyperSpec}(x, na.rm = TRUE, method = c("expSumLog",
  "direct"), ..., label.wavelength)

\S4method{rowQuantiles}{hyperSpec}(x, probs = seq(from = 0, to = 1, by =
  0.25), ..., na.rm = TRUE, label.wavelength, drop = TRUE)

\S4method{rowRanges}{hyperSpec}(x, na.rm = TRUE, ..., label.wavelength)

\S4method{rowRanks}{hyperSpec}(x, ties.method = c("max", "average", "min"),
  ..., na.rm = TRUE, label.wavelength)

\S4method{rowSds}{hyperSpec}(x, ..., na.rm = TRUE, label.wavelength)

\S4method{rowVars}{hyperSpec}(x, na.rm = TRUE, center = NULL, ...,
  label.wavelength)

\method{rowWeightedMeans}{hyperSpec}(x, w = NULL, na.rm = TRUE, ...,
  label.wavelength)

\method{rowWeightedMedians}{hyperSpec}(x, w = NULL, na.rm = TRUE, ...,
  label.wavelength)

\method{rowAvgsPerColSet}{hyperSpec}(X, W = NULL, S, FUN = rowMeans, ...,
  na.rm = TRUE, label.wavelength, tFUN = FALSE)

\method{colAvgsPerRowSet}{hyperSpec}(X, W = NULL, S, FUN = colMeans,
  tFUN = FALSE, ..., na.rm = TRUE, label.spc)
}
\arguments{
\item{x}{hyperSpec object}

\item{label.spc}{labels for the intensity axis for loadings-like statistics}

\item{label.wavelength}{labels for the wavelength axis for scores-like statistics}

\item{drop,na.rm,...}{further parameters to the \link[matrixStats]{matrixStats-package} function}
}
\description{
hyperSpec objects can use matrix functions from package \link[matrixStats]{matrixStats-package}
in addition to the base functions \code{\link[base]{colMeans}}, \code{\link[base]{colSums}},
\code{\link[base]{rowMeans}} and \code{\link[base]{rowSums}}.
}

