% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hyperSpec.matrixStats.unittest}
\alias{hyperSpec.matrixStats.unittest}
\title{hyperSpec.matrixStats unit tests
If \code{\link[svUnit]{svUnit}} is available, run the unit tests and
display the results.}
\usage{
hyperSpec.matrixStats.unittest()
}
\value{
\code{NA} if \code{\link[svUnit]{svUnit}} is not available,
  otherwise \code{TRUE} if all tests are passed successfully. If a test
  fails, \code{hy.unittest} stops with an error.
}
\description{
hyperSpec.matrixStats unit tests
If \code{\link[svUnit]{svUnit}} is available, run the unit tests and
display the results.
}
\examples{
if (require (svUnit)){
 # hyperSpec.matrixStats.unittest ()
}
}
\author{
C. Beleites
}
\seealso{
\code{\link[svUnit]{svUnit}}
}
\keyword{programming}
\keyword{utilities}

